+++ src/main/java/org/mybatis/dynamic/sql/util/SqlProviderAdapter.java
@@ -1,5 +1,5 @@
 /**
- *    Copyright 2016-2017 the original author or authors.
+ *    Copyright 2016-2019 the original author or authors.
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -18,6 +18,7 @@ package org.mybatis.dynamic.sql.util;
 import org.mybatis.dynamic.sql.delete.render.DeleteStatementProvider;
 import org.mybatis.dynamic.sql.insert.render.InsertSelectStatementProvider;
 import org.mybatis.dynamic.sql.insert.render.InsertStatementProvider;
+import org.mybatis.dynamic.sql.insert.render.MultiRowInsertStatementProvider;
 import org.mybatis.dynamic.sql.select.render.SelectStatementProvider;
 import org.mybatis.dynamic.sql.update.render.UpdateStatementProvider;
 
@@ -34,6 +35,10 @@ public class SqlProviderAdapter {
     }
     
     public String insert(InsertStatementProvider<?> insertStatement) {
+        return insertStatement.getInsertStatement();
+    }
+    
+    public String insertMultiple(MultiRowInsertStatementProvider<?> insertStatement) {
         return insertStatement.getInsertStatement();
     }