+++ src/test/java/org/mybatis/dynamic/sql/insert/InsertStatementTest.java
@@ -124,10 +124,6 @@ public class InsertStatementTest {
     @Test
     public void testParallelStream() {
 
-        TestRecord record = new TestRecord();
-        record.setLastName("jones");
-        record.setOccupation("dino driver");
-        
         List<FieldAndValue> mappings = new ArrayList<>();
         
         mappings.add(newFieldAndValue(id.name(), "{record.id}"));
@@ -135,7 +131,7 @@ public class InsertStatementTest {
         mappings.add(newFieldAndValue(lastName.name(), "{record.lastName}"));
         mappings.add(newFieldAndValue(occupation.name(), "{record.occupation}"));
         
-        FieldAndValueCollector<TestRecord> collector = 
+        FieldAndValueCollector collector = 
                 mappings.parallelStream().collect(Collector.of(
                         FieldAndValueCollector::new,
                         FieldAndValueCollector::add,
@@ -151,6 +147,39 @@ public class InsertStatementTest {
     }
     
     private FieldAndValue newFieldAndValue(String fieldName, String valuePhrase) {
+        return FieldAndValue.withFieldName(fieldName)
+                .withValuePhrase(valuePhrase)
+                .build();
+    }
+    
+    @Test
+    public void testParallelStreamForMultiRecord() {
+
+        List<FieldAndValue> mappings = new ArrayList<>();
+        
+        mappings.add(newFieldAndValues(id.name(), "#{records[%s].id}"));
+        mappings.add(newFieldAndValues(firstName.name(), "#{records[%s].firstName}"));
+        mappings.add(newFieldAndValues(lastName.name(), "#{records[%s].lastName}"));
+        mappings.add(newFieldAndValues(occupation.name(), "#{records[%s].occupation}"));
+        
+        FieldAndValueCollector collector = 
+                mappings.parallelStream().collect(Collector.of(
+                        FieldAndValueCollector::new,
+                        FieldAndValueCollector::add,
+                        FieldAndValueCollector::merge));
+                
+        String expectedColumnsPhrase = "(id, first_name, last_name, occupation)";
+        String expectedValuesPhrase = "values"
+                + " (#{records[0].id}, #{records[0].firstName}, #{records[0].lastName}, #{records[0].occupation}),"
+                + " (#{records[1].id}, #{records[1].firstName}, #{records[1].lastName}, #{records[1].occupation})";
+        
+        assertAll(
+                () -> assertThat(collector.columnsPhrase()).isEqualTo(expectedColumnsPhrase),
+                () -> assertThat(collector.multiRowInsertValuesPhrase(2)).isEqualTo(expectedValuesPhrase)
+        );
+    }
+    
+    private FieldAndValue newFieldAndValues(String fieldName, String valuePhrase) {
         return FieldAndValue.withFieldName(fieldName)
                 .withValuePhrase(valuePhrase)
                 .build();